Update generator.yaml to esphome 24.6.0 syntax
ESPHome has changed the config format for OTA updates and one-wire Dallas temp sensors in version 2024.6.0.  It will not accept the old format, so these changes need to be made before the yaml  will compile and firmware installed.
See:
https://esphome.io/changelog/2024.6.0#ota-platforms
https://esphome.io/changelog/2024.6.0#wire

Note that I have not personally tested this specific yaml file, but I have made the same changes to mine and they work correctly.